+++ src/components/SingleContent.tsx
@@ -17,49 +17,42 @@ import {
   getDocs,
   doc,
   onSnapshot,
-  orderBy,
   query,
-  where,
   deleteDoc,
   setDoc
 } from "firebase/firestore"
-import db from "../hooks/firebase"
+import db from "../lib/firebase"
 import { v4 as uuidv4 } from 'uuid';
 import { ITodoData } from '../interfaces/todo'
 import Link from 'next/link';
-import { postsState } from "@atoms/atom"
+import { postsState, userItemState } from "@atoms/atom"
 import { useRecoilState } from 'recoil';
 import { useEffect } from 'react';
 import { useRouter } from 'next/router';
-
-// firebaseのコレクションから複数のドキュメントを取得する
-const q = query(
-  collection(db, 'posts'),
-  where('isDraft', '==', false),
-  where('isTrash', '==', false),
-  orderBy('create')
-)
 
 type props = {
   setTodo: any
 }
 
-export const SingleContent = ({ setTodo }: props) => {
+export const SingleContent = ({ setTodo }:props) => {
   // recoilでatomから取得したグローバルの値
   const [posts, setPosts] = useRecoilState(postsState);
+  const [userItem] = useRecoilState(userItemState);
   const router = useRouter()
-
 
   // recoilで取得した値をonSnapshotで取得、mapで処理を回して全て表示
   // ※onSnapshotを実行すると最初に全てのドキュメントを取得するのでuseEffectで制御
   useEffect(() => {
+    const q = query(
+      collection(db, "users", userItem.uid, 'posts'),
+    )
     const unSub = onSnapshot(q, (querySnapshot) => {
       setPosts(
         querySnapshot.docs.map((post) => ({
           id: post.data().id,
           title: post.data().title,
           text: post.data().text,
-          start: post.data().start,
+          start: new Date(post.data().start).toLocaleDateString(),
           share: post.data().share
         }))
       )
@@ -69,10 +62,10 @@ export const SingleContent = ({ setTodo }: props) => {
 
   useEffect(() => {
     // データベースからデータを取得する
-    const postData = collection(db, "posts");
+    const postData = collection(db, "users", userItem.uid, "posts");
     getDocs(postData).then((snapShot) => {
       setPosts(snapShot.docs.map((doc) => ({ ...doc.data() })));
-      // ...: スプレッド記法、式を複数の要素に展開して、それぞれ関数呼び出す
+      // ...: スプレッド記法、式を複数の要素に展開してそれぞれ関数呼び出す
     });
 
     // リアルタイムで取得
@@ -85,7 +78,7 @@ export const SingleContent = ({ setTodo }: props) => {
   const handleDeletePost = async (targetPost: ITodoData) => {
     setPosts(posts.filter((post: any) => post !== targetPost))
     // postsにfilterをかけてクリックされたpostを抽出
-    await deleteDoc(doc(db, "posts", targetPost.id));
+    await deleteDoc(doc(db, "users", userItem.uid, "posts", targetPost.id));
   }
 
   /**
@@ -96,7 +89,7 @@ export const SingleContent = ({ setTodo }: props) => {
     /**
      * Firebaseに送信する
      */
-    const post = setDoc(doc(db, "posts", postData.id), postData);
+    const post = setDoc(doc(db, "users", userItem.uid, "posts", postData.id), postData);
 
     post.then(() => {
       // inputを空に
@@ -115,10 +108,7 @@ export const SingleContent = ({ setTodo }: props) => {
     const postData: any = {
       id: uuidv4(),
       title: e.target.elements["title"].value,
-      // detail: e.target.elements["detail"].value,
-      // start: e.target.elements["start"].value,
-      // end: e.target.elements["end"].value
-
+      start: e.target.elements["start"].value,
     }
     postAddTask(postData);
   }
