+++ src/lib/firebase.ts
