+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "target": "es5",
+    "target": "es2022",
     "baseUrl": ".",
     "paths": {
       "@*": ["./src/*"]
